requirements.txt
langchain
langchain-community
langgraph
langchain-huggingface
langchain-groq
chromadb
dotenv
pypdf
langchain-chroma
langchain-google-genai
pdf2image
grandalf
langgraph-cli
sentence-transformers
google-adk==0.5.0
yfinance
psutil
litellm
google-generativeai==0.8.5
python-dotenv
google-adk==0.5.0
#google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.32.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.15.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
